Reduce channel metadata concentration and stabilize cleanup gates
The channel registry kept planned and TLON metadata inline, while several
conversation/session/tool helpers remained visible outside the test surface.
This slice moves planned metadata into its own module, merges TLON registry
metadata with its status snapshot owner, and scopes or deletes stale helpers.
It also hardens two flaky local test probes so workspace validation stays
reliable under full-suite load.

Constraint: Keep channel ids, aliases, ordering, and send/serve behavior stable while shrinking registry ownership pressure
Rejected: Broader channel/config ownership rewrite in the same patch | too wide for a behavior-preserving cleanup slice
Rejected: Loosen flaky test assertions only | hides transport timing defects instead of stabilizing the probe path
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep per-channel registry metadata with its snapshot owner; do not re-inline planned or TLON metadata into the root registry file
Tested: cargo fmt --all -- --check; cargo clippy --workspace --all-targets --all-features -- -D warnings; cargo test --workspace --locked; cargo test --workspace --all-features --locked; cargo test -p loongclaw-app channel::registry -- --nocapture; cargo test -p loongclaw-app tlon -- --nocapture; cargo test -p loongclaw-app load_discovery_first_event_summary -- --nocapture; cargo test -p loongclaw-app load_turn_checkpoint_event_summary -- --nocapture; cargo test -p loongclaw-app channel::http::tests::outbound_http_client_does_not_follow_redirects -- --nocapture; cargo test -p loongclaw-daemon browser_companion_diagnostics::tests::collect_browser_companion_diagnostics_retries_transient_probe_timeouts -- --nocapture; scripts/check_dep_graph.sh; LOONGCLAW_ARCH_STRICT=true scripts/check_architecture_boundaries.sh; diff CLAUDE.md AGENTS.md
Not-tested: Full benchmark/smoke suite beyond required CI-parity and targeted regression coverage
